src/user_service_v1/requirements.txt
fastapi
uvicorn 
pymongo
pika
python-dotenv 